refactor(test): extract shared signal collector helper
